gateway_service/app/utils/auth_utils.py
from jose import jwt, JWTError
from fastapi import HTTPException, status
from app.config.config import settings

def get_user_payload_from_token(token: str):
    try:
        payload = jwt.decode(token, settings.JWT_SECRET, algorithms=[settings.JWT_ALGORITHM])
        if not payload.get("user_id"):
            raise HTTPException(status_code=401, detail="Invalid token payload")
        return payload
    except JWTError:
        raise HTTPException(status_code=401, detail="Invalid or expired token")
