gateway_service/requirements.txt
fastapi
uvicorn
httpx
python-dotenv
pydantic
python-jose
pydantic-settings
